+++ DarmuhsTerminalCommands/EventSub/TerminalBeginUsing.cs
@@ -32,11 +32,13 @@ namespace TerminalStuff.EventSub
             if (ConfigSettings.TerminalStuffMain.specialListNum.ContainsKey(Plugin.instance.Terminal.currentNode))
                 return;
 
-            StartofHandling.CheckNetNode(Plugin.instance.Terminal.currentNode);
+            if(!ViewCommands.isVideoPlaying)
+                StartofHandling.CheckNetNode(Plugin.instance.Terminal.currentNode);
         }
 
         internal static void StartUsingTerminalCheck(Terminal instance)
         {
+            TerminalNode nextNode = null;
 
             if (ConfigSettings.TerminalAutoComplete.Value)
             {
@@ -53,7 +55,7 @@ namespace TerminalStuff.EventSub
             //refund init
             if (ConfigSettings.terminalRefund.Value && ConfigSettings.ModNetworking.Value)
             {
-                Plugin.MoreLogs("Syncing items between players for refund command");
+                Plugin.Spam("Syncing items between players for refund command");
                 NetHandler.Instance.SyncDropShipServerRpc();
             }
 
@@ -73,17 +75,24 @@ namespace TerminalStuff.EventSub
             //AlwaysOn Functions
             if (!alwaysOnDisplay)
             {
-                Plugin.Spam("disabling cams views");
+                Plugin.Spam("disabling cams views, alwaysOnDisplay is [DISABLED]");
                 SplitViewChecks.DisableSplitView("neither");
                 ViewCommands.isVideoPlaying = false;
 
                 //Always load to last node or start if alwayson disabled
-                if(lastNode.displayText == "")
+                if(lastNode.displayText == "" || !ConfigSettings.CacheLastTerminalPage.Value)
+                {
                     instance.LoadNewNode(instance.terminalNodes.specialNodes.ToArray()[1]);
+                    nextNode = instance.terminalNodes.specialNodes.ToArray()[1];
+                }
                 else
+                {
                     instance.LoadNewNode(lastNode);
+                    nextNode = lastNode;
+                }
+                    
 
-                if (lastText.Length > 0)
+                if (lastText.Length > 0 && ConfigSettings.CacheLastTerminalPage.Value)
                     LogicHandling.SetTerminalInput(lastText);
             }
             else
@@ -91,53 +100,65 @@ namespace TerminalStuff.EventSub
                 Plugin.MoreLogs("Terminal is Always On, checking for active monitoring to return to.");
                 if (Plugin.instance.isOnMirror || Plugin.instance.isOnCamera || Plugin.instance.isOnMap || Plugin.instance.isOnMiniCams || Plugin.instance.isOnMiniMap || Plugin.instance.isOnOverlay)
                 {
-                    int nodeNum = StartofHandling.FindViewIntByString();
-                    if (nodeNum == -1)
-                        return;
-
-                    TerminalNode returnNode = StartofHandling.FindViewNode(nodeNum);
-                    
-                    if (returnNode == null)
-                        return;
-
-                    instance.LoadNewNode(returnNode);
                     Plugin.Spam($"[returning to camera-type node during AOD]\nMap: {Plugin.instance.isOnMap} \nCams: {Plugin.instance.isOnCamera} \nMiniMap: {Plugin.instance.isOnMiniMap} \nMiniCams: {Plugin.instance.isOnMiniCams} \nOverlay: {Plugin.instance.isOnOverlay}\nMirror: {Plugin.instance.isOnMirror}");
-                    if (lastText.Length > 0)
+                    if (lastText.Length > 0 && ConfigSettings.CacheLastTerminalPage.Value)
                         LogicHandling.SetTerminalInput(lastText);
+
+                    return;
+                }
+                else if(ViewCommands.isVideoPlaying && Plugin.instance.Terminal.currentNode == VideoManager.videoPlayerNode)
+                {
+                    Plugin.Spam($"VideoPlayer: {ViewCommands.isVideoPlaying}\nCurrently Playing: {VideoManager.currentlyPlaying}");
                     return;
                 }
                 else
                 {
-                    Plugin.Spam($"[no matching camera-type nodes during AOD]\nMap: {Plugin.instance.isOnMap} \nCams: {Plugin.instance.isOnCamera} \nMiniMap: {Plugin.instance.isOnMiniMap} \nMiniCams: {Plugin.instance.isOnMiniCams} \nOverlay: {Plugin.instance.isOnOverlay}\nMirror: {Plugin.instance.isOnMirror}");
+                    Plugin.Spam($"[no matching camera-type nodes during AOD]\nMap: {Plugin.instance.isOnMap} \nCams: {Plugin.instance.isOnCamera} \nMiniMap: {Plugin.instance.isOnMiniMap} \nMiniCams: {Plugin.instance.isOnMiniCams} \nOverlay: {Plugin.instance.isOnOverlay}\nMirror: {Plugin.instance.isOnMirror}\nVideoPlayer: {ViewCommands.isVideoPlaying}");
 
-                    if (lastNode == null)
+                    if (ConfigSettings.CacheLastTerminalPage.Value)
                     {
-                        Plugin.WARNING("lastNode terminalNode is null, cached pages failed!");
-                        return;
+                        if (lastNode == null)
+                        {
+                            Plugin.WARNING("lastNode terminalNode is null, cached pages failed!");
+                            return;
+                        }
+
+                        Plugin.Spam($"lastNode Length - {lastNode.displayText.Length}");
+
+                        if (lastNode.displayText.Length < 1)
+                        {
+                            instance.LoadNewNode(instance.terminalNodes.specialNodes.ToArray()[1]);
+                            nextNode = instance.terminalNodes.specialNodes.ToArray()[1];
+                        }
+                            
+                        else if (!Plugin.instance.splitViewCreated && (bool)Plugin.instance.Terminal.displayingPersistentImage)
+                        {
+                            if (lastText.Length > 0)
+                                LogicHandling.SetTerminalInput(lastText);
+                        }
+                        else
+                        {
+                            instance.LoadNewNode(lastNode);
+                            nextNode = lastNode;
+                        }
+                            
                     }
 
-                    Plugin.Spam($"lastNode Length - {lastNode.displayText.Length}");
-
-                    if (lastNode.displayText.Length < 1)
-                        instance.LoadNewNode(instance.terminalNodes.specialNodes.ToArray()[1]);
-                    else if (!Plugin.instance.splitViewCreated && (bool)Plugin.instance.Terminal.displayingPersistentImage)
-                    {
-                        if (lastText.Length > 0)
-                            LogicHandling.SetTerminalInput(lastText);
-                        return;
-                    }  
-                    else
-                        instance.LoadNewNode(lastNode);
-
-                    if (ConfigSettings.networkedNodes.Value)
-                    {
-                        NetHandler.NetNodeReset(true);
-                        NetHandler.Instance.NodeLoadServerRpc(Plugin.instance.Terminal.topRightText.text, Plugin.instance.Terminal.currentNode.name, Plugin.instance.Terminal.currentNode.displayText);
-                    }
-
-                    if (lastText.Length > 0)
+                    if (lastText.Length > 0 && ConfigSettings.CacheLastTerminalPage.Value)
                         LogicHandling.SetTerminalInput(lastText);
                 }
+            }
+
+            if (ConfigSettings.networkedNodes.Value)
+            {
+                if(nextNode == null)
+                {
+                    Plugin.WARNING("Failed to grab nextNode for server sync!");
+                    return;
+                }    
+                Plugin.Spam("sending current node to other users");
+                NetHandler.NetNodeReset(true);
+                NetHandler.Instance.NodeLoadServerRpc(Plugin.instance.Terminal.topRightText.text, nextNode.name, nextNode.displayText);
             }
         }
         //end of void