+++ DarmuhsTerminalCommands/PluginCore/Plugin.cs
@@ -32,7 +32,7 @@ namespace TerminalStuff
         {
             public const string PLUGIN_GUID = "darmuh.TerminalStuff";
             public const string PLUGIN_NAME = "darmuhsTerminalStuff";
-            public const string PLUGIN_VERSION = "3.5.6";
+            public const string PLUGIN_VERSION = "3.5.7";
         }
 
         internal static ManualLogSource Log;